MAGENTO-2109 Working model without guest
